Add TYPE file
readr does a pretty good job of determining types. But sometimes it
makes mistakes. For example, I have a file which contains a column with
comma separated integers. readr ignores the commas, merging the integers
into larger numbers. I need to tell readr to leave them as a character
vector, which I will parse on my own. The hadley way to do this is to
create a `col` object and pass it to the readr::read_tsv function.

I've added a new optional file, TYPE.tsv, which contains a column of
variable names and their types. Rather than have the user enter the
names of parsing functions -- `col_factor(c('A', 'B'))` or
`col_number()`) -- I allow things like `factor(A,B)` and `number`.

Specifically

 * logical     -> col_logical()
 * integer     -> col_integer()
 * double      -> col_double()
 * number      -> col_number()
 * character   -> col_character()
 * factor(A,B) -> col_factor(c('A','B'))
 * factor      -> col_character() # with warning

where A and B are two levels. I currently have no handling for readr
types date, time, datetime.